# docs/05-openapi.yaml
openapi: 3.1.0
info:
  title: TransitGuard API
  version: 1.0.0 (Entwurf, implementierungsreif)
  description: >
    ÖPNV-Companion-API. Meldungs-Ressourcen sind no-store. Auth via X-Device-Token;
    Pro-Endpunkte via Bearer-Entitlement-Token. Vollständige Semantik: docs/04-api.md.
servers:
  - url: https://api.example.de
  - url: https://api.example.de/{basePath}
    variables: { basePath: { default: v1, enum: [v1] } }
security:
  - deviceToken: []
...
paths:
...
  /v1/stops/{stopId}/departures:
    get:
      tags: [schedule]
      parameters:
        - { name: stopId, in: path, required: true, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, default: 20, maximum: 50 } }
      responses:
        "200":
          description: Abfahrten Soll+Ist mit trip_ref
          content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/Departure" } } } }
...
  /v1/cities/{slug}/alerts:
    get:
      tags: [alerts]
      parameters:
        - { name: slug, in: path, required: true, schema: { type: string } }
        - { name: route_id, in: query, schema: { type: string } }
        - { name: stop_id, in: query, schema: { type: string } }
      responses:
        "200": { description: gefilterte Alerts (kein Rauschen), content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/Alert" } } } } }
...
components:
  securitySchemes:
    deviceToken: { type: apiKey, in: header, name: X-Device-Token }
    entitlementToken: { type: http, scheme: bearer, bearerFormat: "et_…" }
...
  schemas:
...
    Departure:
      type: object
      properties:
        trip_ref: { type: object, properties: { trip_id: { type: string }, start_date: { type: string, pattern: '^\\d{8}$' } } }
        route_id: { type: string }
        headsign: { type: string }
        scheduled_time: { type: string, format: date-time }
        estimated_time: { type: [string, "null"], format: date-time }
        delay_s: { type: [integer, "null"] }
        realtime: { type: boolean }
...
    Alert:
      type: object
      properties:
        header: { type: string }
        description: { type: string }
        url: { type: string }
        severity: { type: string, enum: [info, warning, severe] }
        route_refs: { type: array, items: { type: string } }
        until: { type: [string, "null"], format: date-time }
...